remove obstacle stop class
